Created KnowledgeBase
This KnowledgeBase will grow for the newer coders in the group to be able to reference when they want to write code. This is to be as descriptive as possible so when situations arise where a member forgets how something works (which happens to the best of us daily) they can reference this for a working example.